Add subscribe code
We need to create only one gazebo node
in our code, and make it subscribe to several
topics. Unfortunatly, there is no way to make
it generic. We can use the same callback
for several, topics if they use the same message.
    However, there is no garantee that the callbacks
    are going to fire one after the another without
    any racing. So, the best case scenario, is to create
    a callback for each topic.

We might look to have better performance later, by changing
this method. But, we need to find the better one.