Adding a joystick code to allow control the quad with it
This code allow a user to have a control using
the joystick of xbox connected to the USB port
of the machine.
It is going to be integrated later in
the project.
Also needs to remove asio dependency for the keyboard
control
Both of them should be implemented using asynchron future
std library