ALl error has been resolved except SEGFAULT
There is only the gazebo segmentation fault that
stop connecting to gazebo client. Thus we need to find
another way to resolve this issue. Note that, this problem
has emerged since the new upgrade to the new version of gazebo
9.6. It is only this code that is causing the problem. Other
program has no issue. It is related to the thread issue.

The program has been debugged with gdb, valgrind, strace.
It is clean that it is comming from gazebo lib boost asio issues.

Signed-off-by: Omar Shrit <shrit@lri.fr>