Resolve the SEGFAULT issue
The error has been produced by the use of two different versions
of the same boost system library in the same code. Gazebo libraries
had a problem with boost::system::error.

The old version of boost has been disabled. However, now we can not
use Ncurses to control the quadcopter using the keyboard. As it is
has problem with the new version of boost asio. The bug emerged about
one year ago still non resolved by the maintainer of boost asio library

Signed-off-by: Omar Shrit <shrit@lri.fr>