Seems to be working
It seems that mixing smart pointers with thread is not working
correctly, especially shared_ptr, we should aboif as possible this
mix, and replace it with reference as we can.

Gazebo seems to Use shared ptr, and need to use quadrotor class.
We will see, how to improve implementation.

The next step is to make quadrotors close to each others, and recover
data during takeoff and landing and resolve all related errors.

Signed-off-by: Omar Shrit <shrit@lri.fr>